[chore][Feature Request Template] Comment out header descriptions (#9732)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
The descriptions for each header in the feature request template for
this repository are useful to the person filing a request, but serve no
purpose to the issue reader. These can be commented out to in the
displayed markdown to save the user from having to delete each one (or
if not deleted, save the reader from having to parse through extra
information).