[chore] Define scope for 1.0 changes (#10719)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

<!-- Issue number if applicable -->

On the GA roadmap, we defined what modules are in scope, but we did not
explicitly define how to determine the scope within these modules.

This PR adds some wording to define what the scope *within* modules is.
These are roughly the criteria I have personally been using.