[exporterhelper] record metric should log the number of log records before the data are sent to the consumers downstream (#10402)
#### Description
The sender metric within the exporterhelper should measure the number of
items coming into the sender, not what was done with the items
downstream, if the components downstream are mutable. An example of this
is provided as a unit test within this PR.

This PR also addresses nil panics that some users are experiencing. 

#### Link to tracking issue
Fixes https://github.com/open-telemetry/opentelemetry-collector/issues/10033